220719v5

+++ baekjoon/2587.py
@@ -2,5 +2,5 @@ numArray=[]
 for i in range(5):
     numArray+=[int(input())]
 numArray.sort()
-print(sum(numArray)/5)
+print(int(sum(numArray)/5))
 print(numArray[2])
